requirements.txt
flask
flask_bcrypt
flask_login
flask_wtf
wtforms
psycopg2-binary
